requirements.txt
# Python Graph handling package
# https://pypi.org/project/networkx/
networkx
# https://pypi.org/project/pyparsing/
# https://github.com/pyparsing/pyparsing
pyparsing
# https://pypi.org/project/matplotlib/
matplotlib
# https://pypi.org/project/PyYAML/
PyYAML
# https://pypi.org/project/num2words/
num2words
# web server
# https://pypi.org/project/Flask/
Flask
# Mediawiki and SMW API
# https://pypi.org/project/py-3rdparty-mediawiki/
py-3rdparty-mediawiki>=0.0.8
pywikibot>=3.0.20200508
